Fix $this inside "touchstart" always referring to last element
Due to the missing "var" statement a global variable is created.
Since this global variable always contains the last element for
which the "each" function was invoked, the "touchstart" event code
only ever modifies that last element.

See http://jsfiddle.net/5Cqa6/3/ for an example, in which the upper
DIV is broken, and the lower DIV works just fine (since it is the
last).